pyproject.toml
[project]
name = "sql-playgrounds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geoalchemy2>=0.18.0",
    "geopandas>=1.1.1",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=21.0.0",
    "sqlalchemy>=2.0.43",
]
